+++ src/main/java/com/example/main_server/activityLog/slack/team/repository/SlackMemberRepository.java
@@ -1,0 +1,7 @@
+package com.example.main_server.activityLog.slack.team.repository;
+
+import com.example.main_server.activityLog.slack.team.entity.SlackMember;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SlackMemberRepository extends JpaRepository<SlackMember, Long> {
+}
