+++ src/main/java/com/example/main_server/activityLog/teams/teams/chat/repository/TeamsAttachmentRepository.java
@@ -1,0 +1,7 @@
+package com.example.main_server.activityLog.teams.teams.chat.repository;
+
+import com.example.main_server.activityLog.teams.teams.chat.entity.TeamsAttachment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TeamsAttachmentRepository extends JpaRepository<TeamsAttachment, Long> {
+}
