+++ src/main/java/com/example/main_server/activityLog/teams/teams/email/repository/OutlookRecipientRepository.java
@@ -1,0 +1,7 @@
+package com.example.main_server.activityLog.teams.teams.email.repository;
+
+import com.example.main_server.activityLog.teams.teams.email.entity.OutlookRecipient;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OutlookRecipientRepository extends JpaRepository<OutlookRecipient, Long> {
+}
