+++ src/main/java/com/example/main_server/activityLog/teams/teams/onedrive/repository/OnedriveRepository.java
@@ -1,0 +1,7 @@
+package com.example.main_server.activityLog.teams.teams.onedrive.repository;
+
+import com.example.main_server.activityLog.teams.teams.onedrive.entity.Onedrive;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OnedriveRepository extends JpaRepository<Onedrive, Long> {
+}
