--- src/main/java/com/example/main_server/slack/chat/repository/SlackFileRepository.java
@@ -1,7 +1,0 @@
-package com.example.main_server.slack.chat.repository;
-
-import com.example.main_server.slack.chat.entity.SlackFile;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SlackFileRepository extends JpaRepository<SlackFile, Long> {
-}
