--- src/main/java/com/example/main_server/slack/chat/repository/SlackMessageRepository.java
@@ -1,7 +1,0 @@
-package com.example.main_server.slack.chat.repository;
-
-import com.example.main_server.slack.chat.entity.SlackMessage;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SlackMessageRepository extends JpaRepository<SlackMessage, Long> {
-}
