--- src/main/java/com/example/main_server/teams/chat/repository/TeamsAttachmentRepository.java
@@ -1,7 +1,0 @@
-package com.example.main_server.teams.chat.repository;
-
-import com.example.main_server.teams.chat.entity.TeamsAttachment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeamsAttachmentRepository extends JpaRepository<TeamsAttachment, Long> {
-}
