--- src/main/java/com/example/main_server/teams/onedrive/repository/OnedriveRepository.java
@@ -1,7 +1,0 @@
-package com.example.main_server.teams.onedrive.repository;
-
-import com.example.main_server.teams.onedrive.entity.Onedrive;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OnedriveRepository extends JpaRepository<Onedrive, Long> {
-}
